Move zarr archive ingest into task
