+++ dandiapi/api/admin.py
@@ -1,6 +1,6 @@
 import csv
 
-from django.contrib import admin
+from django.contrib import admin, messages
 from django.contrib.admin.options import TabularInline
 from django.contrib.auth.admin import UserAdmin as BaseUserAdmin
 from django.contrib.auth.models import User
@@ -11,9 +11,9 @@ from django.http.request import HttpRequest
 from django.http.response import HttpResponse
 from django.urls import reverse
 from django.utils.safestring import mark_safe
+from django.utils.translation import ngettext
 from guardian.admin import GuardedModelAdmin
 
-from dandiapi.api.management.commands.ingest_zarr_archive import ingest_zarr_archive
 from dandiapi.api.models import (
     Asset,
     AssetBlob,
@@ -26,6 +26,7 @@ from dandiapi.api.models import (
     ZarrArchive,
     ZarrUploadFile,
 )
+from dandiapi.api.tasks.zarr import ingest_zarr_archive
 from dandiapi.api.views.users import social_account_to_dict
 
 admin.site.site_header = 'DANDI Admin'
@@ -162,7 +163,19 @@ class ZarrArchiveAdmin(admin.ModelAdmin):
     def ingest_zarr_archive(self, request, queryset):
         for zarr in queryset:
             zarr: ZarrArchive
-            ingest_zarr_archive(str(zarr.id))
+            ingest_zarr_archive.delay(zarr_id=(str(zarr.id)))
+
+        # Return message
+        self.message_user(
+            request,
+            ngettext(
+                '%d zarr archive has begun ingesting.',
+                '%d zarr archives have begun ingesting.',
+                queryset.count(),
+            )
+            % queryset.count(),
+            messages.SUCCESS,
+        )
 
     def __init__(self, model, admin_site) -> None:
         super().__init__(model, admin_site)
